# proto/gen/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/instance/profile:
        get:
            tags:
                - InstanceService
            description: Gets the instance profile.
            operationId: InstanceService_GetInstanceProfile
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/InstanceProfile'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/instance/settings/notification:testEmail:
        post:
            tags:
                - InstanceService
            description: Tests notification email delivery with the provided or stored SMTP settings.
            operationId: InstanceService_TestInstanceEmailSetting
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TestInstanceEmailSettingRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        InstanceProfile:
            type: object
            properties:
                version:
                    type: string
                    description: Version is the current version of instance.
                demo:
                    type: boolean
                    description: Demo indicates if the instance is in demo mode.
                instanceUrl:
                    type: string
                    description: Instance URL is the URL of the instance.
                admin:
                    allOf:
                        - $ref: '#/components/schemas/User'
                    description: |-
                        The first administrator who set up this instance.
                         When null, instance requires initial setup (creating the first admin account).
                commit:
                    type: string
                    description: Commit is the current build commit of instance.
            description: Instance profile message containing basic instance information.
...
        NotificationSetting_EmailSetting:
            type: object
            properties:
                enabled:
                    type: boolean
                smtpHost:
                    type: string
                smtpPort:
                    type: integer
                    format: int32
                smtpUsername:
                    type: string
                smtpPassword:
                    writeOnly: true
                    type: string
                fromEmail:
                    type: string
                fromName:
                    type: string
                replyTo:
                    type: string
                useTls:
                    type: boolean
                useSsl:
                    type: boolean
            description: Email delivery configuration for notifications.
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...
        TestInstanceEmailSettingRequest:
            type: object
            properties:
                email:
                    allOf:
                        - $ref: '#/components/schemas/NotificationSetting_EmailSetting'
                    description: Optional. SMTP email settings to test. If omitted, the stored notification email setting is used.
                recipientEmail:
                    type: string
                    description: Optional. Recipient email address. If omitted, the current user's email address is used.
            description: Request message for TestInstanceEmailSetting method.
...
        User:
            required:
                - role
                - username
                - state
            type: object
            properties:
                name:
                    type: string
                    description: |-
                        The resource name of the user.
                         Format: users/{user}
                role:
                    enum:
                        - ROLE_UNSPECIFIED
                        - ADMIN
                        - USER
                    type: string
                    description: The role of the user.
                    format: enum
                username:
                    type: string
                    description: Required. The unique username for login.
                email:
                    type: string
                    description: Optional. The email address of the user.
                displayName:
                    type: string
                    description: Optional. The display name of the user.
                avatarUrl:
                    type: string
                    description: Optional. The avatar URL of the user.
                description:
                    type: string
                    description: Optional. The description of the user.
                password:
                    writeOnly: true
                    type: string
                    description: Input only. The password for the user.
                state:
                    enum:
                        - STATE_UNSPECIFIED
                        - NORMAL
                        - ARCHIVED
                    type: string
                    description: The state of the user.
                    format: enum
                createTime:
                    readOnly: true
                    type: string
                    description: Output only. The creation timestamp.
                    format: date-time
                updateTime:
                    readOnly: true
                    type: string
                    description: Output only. The last update timestamp.
                    format: date-time
...